cCheck code
I wrote the code to check for a collision and change a variable based on
the result. If checker=1 we'll run collision, if checker=0, we won't. I
think we'll do this in main() right? The alternative is just running
collision() within cCheck() which I can do if you want. I also fixed a
typo I made,